fix to naut type conversion: err if unknown, handle more types
Previously, an unhandled type was silently treated as void, which could
result in compile errors. Now this results in compile time errors due to
static_assert if supported, or in runtime errors

Also adds previously missing types.

We add both (unsigned) long and long long as (u)int64_t is defined as
long long on macos and as long on unix.
Thus we have to do extra checks for both to have them supported.

We also assert that type is actually 64 bit as long is 32 bit on
windows, which could cause a bug if windows were to be supported.